done with lab2

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,9 +6,9 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
-        return result;
+        double quotient = (double)top / bottom;
+        long result = Math.round(quotient);
+        return (int)result;
     }
 
     public static void main(String[] args) {
